fixed target graphics bug

+++ candidbot_controller.py
@@ -64,8 +64,8 @@ class CandidbotController:
       # Draw green circle on target
       if self.target:
         left, top, width, height = self.target.box
-        center_x = left + (width / 2)
-        center_y = top + (height / 2)
+        center_x = int(left + (width / 2))
+        center_y = (top + (height / 2))
         cv2.circle(frame, (center_x, center_y), 20, (0, 255, 0), -1)
 
       cv2.imshow(self.window_name, frame)
